Email Workflow v1/V2/requirements.txt
openai-agents>=0.0.17
pydantic>=2.7.0
python-dotenv>=1.0.1
google-api-python-client>=2.170.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
streamlit>=1.37.0
